@@ -1,17 +1,17 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="IPAddressRange" Version="6.0.0" />
-    <PackageReference Include="GraphQL.Client" Version="6.0.0" />
-    <PackageReference Include="GraphQL.Client.Serializer.Newtonsoft" Version="6.0.0" />
-    <PackageReference Include="GraphQL.Client.Serializer.SystemTextJson" Version="6.0.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Components" Version="6.0.25" />
+    <PackageReference Include="GraphQL.Client" Version="6.0.3" />
+    <PackageReference Include="GraphQL.Client.Serializer.Newtonsoft" Version="6.0.3" />
+    <PackageReference Include="GraphQL.Client.Serializer.SystemTextJson" Version="6.0.3" />
+    <PackageReference Include="Microsoft.AspNetCore.Components" Version="8.0.4" />
   </ItemGroup>
 
   <ItemGroup>
