@@ -1,4 +1,5 @@
 ﻿using NUnit.Framework;
+using NUnit.Framework.Legacy;
 using FWO.GlobalConstants;
 using FWO.Api.Data;
 
@@ -20,31 +21,31 @@ namespace FWO.Test
         [Test]
         public void TestDistName()
         {
-            Assert.AreEqual("", dn1.UserName);
-            Assert.AreEqual("", dn1.Role);
-            Assert.AreEqual("", dn1.Group);
-            Assert.AreEqual(0, dn1.Root.Count);
-            Assert.AreEqual(0, dn1.Path.Count);
-            Assert.AreEqual("", dn1.getTenant());
-            Assert.AreEqual(false, dn1.IsInternal());
+            ClassicAssert.AreEqual("", dn1.UserName);
+            ClassicAssert.AreEqual("", dn1.Role);
+            ClassicAssert.AreEqual("", dn1.Group);
+            ClassicAssert.AreEqual(0, dn1.Root.Count);
+            ClassicAssert.AreEqual(0, dn1.Path.Count);
+            ClassicAssert.AreEqual("", dn1.GetTenantNameViaLdapTenantLevel());
+            ClassicAssert.AreEqual(false, dn1.IsInternal());
 
-            Assert.AreEqual("intuser2", dn2.UserName);
-            Assert.AreEqual("", dn2.Role);
-            Assert.AreEqual("", dn2.Group);
-            Assert.AreEqual(2, dn2.Root.Count);
-            Assert.AreEqual(4, dn2.Path.Count);
-            Assert.AreEqual("tenant2", dn2.getTenant(3));
-            Assert.AreEqual(true, dn2.IsInternal());
+            ClassicAssert.AreEqual("intuser2", dn2.UserName);
+            ClassicAssert.AreEqual("", dn2.Role);
+            ClassicAssert.AreEqual("", dn2.Group);
+            ClassicAssert.AreEqual(2, dn2.Root.Count);
+            ClassicAssert.AreEqual(4, dn2.Path.Count);
+            ClassicAssert.AreEqual("tenant2", dn2.GetTenantNameViaLdapTenantLevel(3));
+            ClassicAssert.AreEqual(true, dn2.IsInternal());
 
-            Assert.AreEqual("usergroup3", dn3.UserName);
-            Assert.AreEqual("usergroup3", dn3.Role);
-            Assert.AreEqual("usergroup3", dn3.Group);
-            Assert.AreEqual(2, dn3.Root.Count);
-            Assert.AreEqual("somewhere", dn3.Root[0]);
-            Assert.AreEqual(3, dn3.Path.Count);
-            Assert.AreEqual("groups", dn3.Path[0]);
-            Assert.AreEqual("", dn3.getTenant(0));
-            Assert.AreEqual(false, dn3.IsInternal());
+            ClassicAssert.AreEqual("usergroup3", dn3.UserName);
+            ClassicAssert.AreEqual("usergroup3", dn3.Role);
+            ClassicAssert.AreEqual("usergroup3", dn3.Group);
+            ClassicAssert.AreEqual(2, dn3.Root.Count);
+            ClassicAssert.AreEqual("somewhere", dn3.Root[0]);
+            ClassicAssert.AreEqual(3, dn3.Path.Count);
+            ClassicAssert.AreEqual("groups", dn3.Path[0]);
+            ClassicAssert.AreEqual("", dn3.GetTenantNameViaLdapTenantLevel(0));
+            ClassicAssert.AreEqual(false, dn3.IsInternal());
         }
     }
 }
