@@ -2,6 +2,7 @@
 using FWO.Report.Filter.Ast;
 using FWO.Report.Filter.Exceptions;
 using NUnit.Framework;
+using NUnit.Framework.Legacy;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -114,7 +115,7 @@ namespace FWO.Test
             }
             catch (SyntaxException exception)
             {
-                Assert.AreEqual("No token but one was expected", exception.Message);
+                ClassicAssert.AreEqual("No token but one was expected", exception.Message);
             }
         }
 
