@@ -1,4 +1,5 @@
 ﻿using NUnit.Framework;
+using NUnit.Framework.Legacy;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -58,7 +59,7 @@ namespace FWO.Test
               bw.Write(pdf);
             }
             Assert.That(filePath, Does.Exist);
-            Assert.Greater(new System.IO.FileInfo(filePath).Length, 5000);
+            ClassicAssert.Greater(new System.IO.FileInfo(filePath).Length, 5000);
         }
 
         [OneTimeTearDown]
